requirements.txt
dataclasses==0.7
natsort==6.2.0
numpy==1.17.4
Pillow==6.2.1
pyparsing==2.3.0
PyYAML==5.2
remi==2019.9
requests==2.22.0
tflite-runtime==1.14.0